Move StoreUtil to enterprise ha store package
